Download the human gene nomenclature dataset

download_hgnc_dataset() gets the latest HGNC approved data set.

Usage

download_hgnc_dataset(
  url = latest_archive_url(),
  path = getwd(),
  filename = basename(url)
)

Arguments

url

A character string naming the URL of the HGNC dataset. It defaults to the latest available archive.

path

The directory path where the downloaded file is to be saved. By default, this value is NULL and the data is imported directly into memory without saving into disk.

filename

A character string with the name of the saved file. By default, this is inferred from the last part of the URL.

Value

The path to the saved file.