hgnc: Import Human Gene Nomenclature

A set of routines to quickly download and import the HUGO Gene Nomenclature Committee (HGNC) data set on mapping of gene symbols to gene entries in other genomic databases or resources.

See also

Author

Maintainer: Ramiro Magno rmagno@pattern.institute (ORCID)

Authors:

Isabel Duarte iduarte@pattern.institute (ORCID)

Jacob Munro jacob.e.munro@gmail.com (ORCID)

Other contributors:

Ana-Teresa Maia maia.anateresa@gmail.com (ORCID) [contributor]

Pattern Institute (04jrgd746) [copyright holder, funder]