List monthly or quarterly archives

This function lists the monthly and quarterly archives currently available from the HGNC's Google Cloud Storage.

Usage

list_archives(release = c("monthly", "quarterly"))

Arguments

release

Series type: "monthly" or "quarterly".

Value

A tibble of available archives for download with the following columns:

series: whether "monthly" or "quarterly".

dataset: type of data set ("hgnc_complete_set", "symbol-changes-monthly", "withdrawn" or "symbol-changes-quarterly").

file: file name.

date: update date.

size: file size.

last_modified: date-time of file last modification on server.

md5sum: MD5 checksum.

url: direct address to the archive.

Examples

try(list_archives())
#> # A tibble: 135 × 9
#>    series  dataset     archive file  date       size  last_modified       md5sum
#>    <chr>   <chr>       <chr>   <chr> <date>     <chr> <dttm>              <chr> 
#>  1 monthly hgnc_compl… hgnc_c… hgnc… 2021-03-01 15.2… 2024-10-01 14:52:08 5bdr4…
#>  2 monthly hgnc_compl… hgnc_c… hgnc… 2021-04-01 15.2… 2024-10-01 14:52:02 kQTSM…
#>  3 monthly hgnc_compl… hgnc_c… hgnc… 2021-05-01 15.2… 2024-10-01 14:52:15 dRPRI…
#>  4 monthly hgnc_compl… hgnc_c… hgnc… 2021-06-01 15.8… 2024-10-01 14:51:59 GzZd4…
#>  5 monthly hgnc_compl… hgnc_c… hgnc… 2021-07-01 15.8… 2024-10-01 14:52:10 POYCr…
#>  6 monthly hgnc_compl… hgnc_c… hgnc… 2021-08-01 15.9… 2024-10-01 14:52:01 rsxim…
#>  7 monthly hgnc_compl… hgnc_c… hgnc… 2021-09-01 15.9… 2024-10-01 14:52:14 UH0Q+…
#>  8 monthly hgnc_compl… hgnc_c… hgnc… 2021-10-01 15.9… 2024-10-01 14:51:55 0m9EC…
#>  9 monthly hgnc_compl… hgnc_c… hgnc… 2021-11-01 15.9… 2024-10-01 14:52:08 xihTt…
#> 10 monthly hgnc_compl… hgnc_c… hgnc… 2021-12-01 16.0… 2024-10-01 14:51:53 jbFMg…
#> # ℹ 125 more rows
#> # ℹ 1 more variable: url <chr>