Filter genes by keyword

Filter the HGNC data set by a keyword to be looked up in the columns containing gene names or symbols. By default, it will look up in symbol, name, alias_symbol, alias_name, prev_symbol and prev_name. Note that this function in dive into list-columns and match return a hit result if at least one of the strings matches the keyword.

Usage

filter_by_keyword(
  tbl,
  keyword,
  cols = c("symbol", "name", "alias_symbol", "alias_name", "prev_symbol", "prev_name")
)

Arguments

tbl

A tibble containing the HGNC data set, typically obtained with import_hgnc_dataset().

keyword

A keyword or a regular expression to be used as search criterion.

cols

Columns to be looked up.

Value

A tibble of the HGNC data set filtered by observations matching the keyword.

Examples

# Start by retrieving the HGNC data set
hgnc_tbl <- import_hgnc_dataset()

# Search for entries containing "TP53" in the HGNC data set
hgnc_tbl %>%
  filter_by_keyword('TP53') %>%
  dplyr::select(1:4)
#> # A tibble: 47 × 4
#>    hgnc_id    hgnc_id2 symbol     name                                          
#>    <chr>      <chr>    <chr>      <chr>                                         
#>  1 HGNC:49685 49685    ABHD15-AS1 ABHD15 antisense RNA 1                        
#>  2 HGNC:20679 20679    ANO9       anoctamin 9                                   
#>  3 HGNC:13276 13276    EI24       EI24 autophagy associated transmembrane prote…
#>  4 HGNC:3345  3345     ENC1       ectodermal-neural cortex 1                    
#>  5 HGNC:4136  4136     GAMT       guanidinoacetate N-methyltransferase          
#>  6 HGNC:6568  6568     LGALS7     galectin 7                                    
#>  7 HGNC:53222 53222    LINC02303  long intergenic non-protein coding RNA 2303   
#>  8 HGNC:16841 16841    LITAF      lipopolysaccharide induced TNF factor         
#>  9 HGNC:6762  6762     MAD1L1     mitotic arrest deficient 1 like 1             
#> 10 HGNC:17637 17637    PERP       p53 apoptosis effector related to PMP22       
#> # … with 37 more rows

# The same as above but restrict the search to the `symbol` column
hgnc_tbl %>%
  filter_by_keyword('TP53', cols = 'symbol') %>%
  dplyr::select(1:4)
#> # A tibble: 23 × 4
#>    hgnc_id    hgnc_id2 symbol    name                                           
#>    <chr>      <chr>    <chr>     <chr>                                          
#>  1 HGNC:11998 11998    TP53      tumor protein p53                              
#>  2 HGNC:29984 29984    TP53AIP1  tumor protein p53 regulated apoptosis inducing…
#>  3 HGNC:11999 11999    TP53BP1   tumor protein p53 binding protein 1            
#>  4 HGNC:12000 12000    TP53BP2   tumor protein p53 binding protein 2            
#>  5 HGNC:16328 16328    TP53BP2P1 tumor protein p53 binding protein 2 pseudogene…
#>  6 HGNC:43652 43652    TP53COR1  tumor protein p53 pathway corepressor 1        
#>  7 HGNC:19373 19373    TP53I3    tumor protein p53 inducible protein 3          
#>  8 HGNC:16842 16842    TP53I11   tumor protein p53 inducible protein 11         
#>  9 HGNC:25102 25102    TP53I13   tumor protein p53 inducible protein 13         
#> 10 HGNC:18022 18022    TP53INP1  tumor protein p53 inducible nuclear protein 1  
#> # … with 13 more rows

# Match "TP53" exactly in the `symbol` column
hgnc_tbl %>%
  filter_by_keyword('^TP53$', cols = 'symbol') %>%
  dplyr::select(1:4)
#> # A tibble: 1 × 4
#>   hgnc_id    hgnc_id2 symbol name             
#>   <chr>      <chr>    <chr>  <chr>            
#> 1 HGNC:11998 11998    TP53   tumor protein p53

# `filter_by_keyword()` is vectorised over `keyword`
hgnc_tbl %>%
  filter_by_keyword(c('^TP53$', '^PIK3CA$'), cols = 'symbol') %>%
  dplyr::select(1:4)
#> # A tibble: 2 × 4
#>   hgnc_id    hgnc_id2 symbol name                                               
#>   <chr>      <chr>    <chr>  <chr>                                              
#> 1 HGNC:11998 11998    TP53   tumor protein p53                                  
#> 2 HGNC:8975  8975     PIK3CA phosphatidylinositol-4,5-bisphosphate 3-kinase cat…